Add provider for managing current user state
Create CurrentUserContext and SetCurrentUserContext to provide access to the current user and the function to set the current user.
Implement custom hooks useCurrentUser and useSetCurrentUser for convenient access to the current user and the function to set the current user.
Add CurrentUserProvider component to encapsulate the state management of the current user.
Fetch the current user data from the server on component mount using an API request to dj-rest-auth/user/.
Implement Axios interceptors to handle authentication tokens:
Intercept requests using axiosReq and attempt to refresh the authentication token. If the token refresh fails, navigate to the login page.
Intercept responses using axiosRes and handle token expiration errors by attempting to refresh the token. If the token refresh fails, navigate to the login page.
Provide the current user and set current user functions to child components using the context providers.