Add 404 error component
Added a component that shows inplaced of a page when the user tries to navigate to a post, trainer, or pokemon that does not exist. A new state variable noResults is set if the API gets a 404 error response. If noResults is true the Error404 component is returned instead with a message reflecting the error.