Add notifcation contexts
Added a new context. useSetCurrentNotification can be called to set the notification message, and useCurrentNotification is called by Notification.js to display the message as a disappearing pop-up in the top right corner. Notification.js imported into App.js. NotificationContext.js is imported into index.js as a parent of App.js