Fixed infinite scroll hasMore behaviour
Set a max-height to the div that shows the pokemon collection just to make the page easier to navigate. Assigned it as the scrollable target to fix the issue I had with loading more data. Used an if statement to set has more to false if the size of the collection is smaller than the size initially loaded.